Fixed Saving and Loading
Now you can fully save your dorm. When loading it recreates dorms from the dorm.lvl file that is saved in Application.persistentDataPath. It saves the game when the game is paused, or you exit the application. It is still not very safe against erros, but fine for now. It also saves build floors and rooms.